Move MonthDataStorage to api project
