+++ api/src/main/java/org/itsallcode/whiterabbit/api/MonthDataStorage.java
@@ -1,0 +1,18 @@
+package org.itsallcode.whiterabbit.api;
+
+import java.time.YearMonth;
+import java.util.List;
+import java.util.Optional;
+
+import org.itsallcode.whiterabbit.api.model.MonthData;
+
+public interface MonthDataStorage
+{
+    Optional<MonthData> load(YearMonth date);
+
+    void store(YearMonth yearMonth, MonthData record);
+
+    List<YearMonth> getAvailableDataMonths();
+
+    List<MonthData> loadAll();
+}
